backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
anthropic>=0.40
pydantic>=2.7
aiosqlite>=0.20
python-dotenv>=1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0
